created helper file and new files
